APP-1117 APP-1104

Closes APP-1117

See merge request mobile/ktobet-asia-ios!113

+++ ktobet-asia-ios/domain/api/PlayerApi.swift
@@ -51,12 +51,12 @@ class PlayerApi {
         }
     }
     
-    func getCashBalance() -> Single<ResponseData<Int>> {
+    func getCashBalance() -> Single<ResponseData<Double>> {
         let target = APITarget(baseUrl: httpClient.baseUrl,
                                path: "api/cash/balance",
                                method: .get,
                                task: .requestPlain,
                                header: httpClient.headers)
-        return httpClient.request(target).map(ResponseData<Int>.self)
+        return httpClient.request(target).map(ResponseData<Double>.self)
     }
 }
